% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{aa_summary_by_user}
\alias{aa_summary_by_user}
\title{aa_summary_by_user}
\usage{
aa_summary_by_user(aa, max_time_interval = "2 days", by_week = FALSE)
}
\arguments{
\item{aa}{dataframe of activity data. Must include column names corresponding to the other params}

\item{max_time_interval}{The cutoff for duration of any session (duration data type, accepts recognised strings). Any session registering above this is filtered out as it is likely to be someone just logging in and then walking away. Defaults to 2 days.}

\item{by_week}{Boolean. TRUE if 'week' field is to be grouped by as well.}
}
\description{
This summarises 'activity' data, according to the fields:
id : A unique key for the user
session : A unique key for each 'session', which denotes a single login from a single user
timestamp : When the event occured
}
\details{
The defaults have been setup to work with Blackboard DDA report data, hence some of the strange names.
}
\keyword{aa}
\keyword{accumulator}
\keyword{activity}
\keyword{log}
\keyword{summary}
\keyword{user}
