% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity_summary}
\alias{activity_summary}
\title{activity_summary}
\usage{
activity_summary(aa, content = c("event", "handle", "data",
  "content_pk1"), max_time_interval = "2 days")
}
\arguments{
\item{aa}{dataframe of activity data. Must include column names corresponding to the other params}

\item{content}{A character vector of the field names that define the content}

\item{max_time_interval}{The cutoff for duration of any session (duration data type, accepts recognised strings). Any session registering above this is filtered out as it is likely to be someone just logging in and then walking away. Defaults to 2 days.}
}
\description{
This summarises 'activity' data, according to the fields:
id : A unique key for the user
session : A unique key for each 'session', which denotes a single login from a single user
timestamp : When the event occured
content : A collection of field names that describe what was accessed
}
\details{
The defaults have been setup to work with Blackboard DDA report data, hence some of the strange names.
}
\keyword{aa}
\keyword{accumulator}
\keyword{activity}
\keyword{log}
\keyword{summary}
