% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{as_matrix_greedily}
\alias{as_matrix_greedily}
\title{as_matrix_greedily}
\usage{
as_matrix_greedily(df)
}
\arguments{
\item{df}{data frame to be coerced into numeric}
}
\description{
Takes a data frame and tries to return a numeric matrix, first
converting as many fields that might be possible numerics first.
This includes 'duration' type characters of the form 31s. Others
to be included as needed
}
\keyword{coerce}
\keyword{dataframe}
\keyword{matrix}
\keyword{numeric}
