% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{best_bet}
\alias{best_bet}
\title{best_bet}
\usage{
best_bet(candidates)
}
\arguments{
\item{candidates}{list of values to choose from}
}
\description{
Function that takes in a list of values and attempts to
return a 'best bet' of what the value should be. Can handle
various data types.
}
\keyword{best}
\keyword{bet}
\keyword{filter}
\keyword{na}
\keyword{remove}
