% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorers.R
\name{heatmap_lakit}
\alias{heatmap_lakit}
\title{heatmap_lakit}
\usage{
heatmap_lakit(x, cluster_method = "average",
  distance_metric = "euclidian", k = 1, ...)
}
\arguments{
\item{x}{numeric matrix of the values to be plotted}

\item{cluster_method}{method of clustering}

\item{distance_metric}{method of measuring distance between two points}

\item{...}{other parameters passed to heatmap}
}
\description{
Reworked heatmap with different defaults and easier choosing
of methods
}
\keyword{clustering}
\keyword{heatmap}
