% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{is_duration_string}
\alias{is_duration_string}
\title{is_duration_string}
\usage{
is_duration_string(x)
}
\arguments{
\item{x}{the string to be examined}
}
\description{
Takes a character string and attempts to determine if
it could be a duration type string.
Will return true if string starts with a number followed by
an 's', and is also possibly followed by a space and then an
open parenthesis and a ~ and a number and then anything.
So it will falsely return "31.2s (~40 cubits)" as TRUE
}
\keyword{detect}
\keyword{duration}
