% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{is_numericable}
\alias{is_numericable}
\title{is_numericable}
\usage{
is_numericable(x)
}
\arguments{
\item{x}{dubious vector}
}
\description{
Checks to see if a vector can be coerced to numeric
}
\keyword{coerce}
\keyword{numeric}
