% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{namdiff}
\alias{namdiff}
\title{names difference}
\usage{
namdiff(x, y)
}
\arguments{
\item{x}{a data frame}

\item{y}{a data frame}
}
\value{
a character vector
}
\description{
Looks at what names in one data frame are
not in the names of another data frame
}
\examples{

x <- data.frame(A = 1:3, B = 9:11)
y <- data.frame(A = 1:10, C = 11:20)
namdiff(x, y)
}
