% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{pagebreak}
\alias{pagebreak}
\title{pagebreak}
\usage{
pagebreak()
}
\description{
A function that includes a pagebreak in rmarkdown,
in the correct format depending on the chosen output
of the document (pdf or html)
}
\details{
To include in a markdown document use:
`r pagebreak()`
}
\concept{formatters}
