% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{plot_timestamp_spectrum}
\alias{plot_timestamp_spectrum}
\title{plot_timestamp_spectrum}
\usage{
plot_timestamp_spectrum(timelist_differences, trans = "log10", ...)
}
\arguments{
\item{trans}{x-axis transformation, defaults to 'log10'}

\item{...}{Extra parameters passed to the geom_density geom}

\item{timelist_difference}{a vector of durations (see lubridate) as outputted by the timelist_to_difference function}
}
\description{
This takes the output from timelist_to_difference and generates
a frequency density plot of the differences, which is akin to
a freqency ~ amplitude plot used in spectral analysis.
}
\keyword{frequency}
\keyword{histogram}
\keyword{plot}
\keyword{spectrum}
\keyword{time}
