% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{str_name_to_first_last}
\alias{str_name_to_first_last}
\title{str_name_to_first_last}
\usage{
str_name_to_first_last(x)
}
\arguments{
\item{x}{a name string}
}
\description{
Switches Lastname, Firstname to Firstname Lastname or back.
Uses presence of "," to determine which way to go
}
\examples{

switch_name_style("Teresa May")
switch_name_style("Trump, Donald")

}
