% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{str_name_to_last_first}
\alias{str_name_to_last_first}
\title{str name to last first}
\usage{
str_name_to_last_first(x, sep = ", ")
}
\arguments{
\item{x}{string of names}
}
\value{
string vector
}
\description{
Takes a name in the format Firstname Lastname
and converts to Lastname, Firstname
}
