% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{str_name_trim_pref}
\alias{str_name_trim_pref}
\title{str name trim preferred}
\usage{
str_name_trim_pref(x)
}
\arguments{
\item{x}{a string}
}
\value{
string vector
}
\description{
Removes any text within ()
Used to remove preferred names in name data, e.g.
Robert (Bob) Johnson should become Robert Johnson
The space from before the ( is removed as well.
}
\examples{
str_name_trim_pref("Robert (Bob) Johnson")
}
