% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{switch_name_style}
\alias{switch_name_style}
\title{switch name style}
\usage{
switch_name_style(x)
}
\arguments{
\item{x}{a name string}
}
\description{
Switches Lastname, Firstname to Firstname Lastname or back.
Uses presence of "," to determine which way to go
}
\examples{

switch_name_style("Teresa May")
switch_name_style("Trump, Donald")

}
