% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{to_duration_from_string}
\alias{to_duration_from_string}
\title{to_duration_from_string}
\usage{
to_duration_from_string(x)
}
\arguments{
\item{x}{a string, of form "310s (~5.17 minutes)" form}
}
\description{
Gets strings that were durations and returns a duration
}
\keyword{character}
\keyword{clean}
\keyword{duration}
\keyword{tidy}
