% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{timelist_to_difference}
\alias{timelist_to_difference}
\title{timelist_to_difference function}
\usage{
timelist_to_difference(timestamps, ...)
}
\arguments{
\item{timestamps}{a vector of POSIXct}
}
\description{
This function converts a vector of timestamps (POSIXct/lt)
into a collection of all the differences between each timestamp
with each other. The idea is to head towards a frequency distribution
of the time series, but rather than using a signal it is using a
sequences of events or impulses
}
\keyword{activity}
\keyword{timestamp}
