% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{timelist_to_difference_assembler}
\alias{timelist_to_difference_assembler}
\title{timelist_to_difference_assembler}
\usage{
timelist_to_difference_assembler(timestamps, ids)
}
\arguments{
\item{timestamps}{a vector of timestamp}

\item{ids}{a corresponding vector of ids, matching to the timestamps}
}
\description{
timelist_to_difference_assembler
}
\examples{
aa <- data.frame(ids = rep(c(1,2,3,4),25), ts = as.POSIXlt.POSIXct(sample(10000000:11100000 , 100, replace = FALSE)))

}
\seealso{
Other timelist: \code{\link{timelist_to_difference}}
}
\concept{timelist}
\keyword{activity}
\keyword{timestamp}
